refactor: separates contribution and ticket donation done pages
